+++ src/test/kotlin/com/atlassian/performance/tools/virtualusers/measure/ClusterNodeCounterTest.kt
@@ -1,5 +1,6 @@
 package com.atlassian.performance.tools.virtualusers.measure
 
+import com.atlassian.performance.tools.virtualusers.diagnostics.DisabledDiagnostics
 import org.assertj.core.api.Assertions
 import org.junit.Test
 import java.util.function.Supplier
@@ -10,9 +11,12 @@ class ClusterNodeCounterTest {
     fun shouldCountAFailingNode() {
         val clusterNodeCounter = ClusterNodeCounter()
 
-        clusterNodeCounter.count(Supplier {
-            throw Exception("Failed to identify jira Node")
-        })
+        clusterNodeCounter.count(
+            Supplier {
+                throw Exception("Failed to identify jira Node")
+            },
+            DisabledDiagnostics()
+        )
 
         val nodeIdentifyBuilder = StringBuilder()
         clusterNodeCounter.dump(nodeIdentifyBuilder)